Use 64-bit unsigned integers for timestamps

+++ mongoose.h
@@ -407,21 +407,21 @@ void mg_log_set_callback(void (*fn)(const void *, int, void *), void *param);
 
 
 struct mg_timer {
-  int period_ms;            // Timer period in milliseconds
-  int flags;                // Possible flags values below
-#define MG_TIMER_REPEAT 1   // Call function periodically, otherwise run once
-#define MG_TIMER_RUN_NOW 2  // Call immediately when timer is set
-  void (*fn)(void *);       // Function to call
-  void *arg;                // Function agrument
-  unsigned long expire;     // Expiration timestamp in milliseconds
-  struct mg_timer *next;    // Linkage in g_timers list
+  int period_ms;             // Timer period in milliseconds
+  int flags;                 // Possible flags values below
+#define MG_TIMER_REPEAT 1    // Call function periodically, otherwise run once
+#define MG_TIMER_RUN_NOW 2   // Call immediately when timer is set
+  void (*fn)(void *);        // Function to call
+  void *arg;                 // Function agrument
+  unsigned long long expire; // Expiration timestamp in milliseconds
+  struct mg_timer *next;     // Linkage in g_timers list
 };
 
 extern struct mg_timer *g_timers;  // Global list of timers
 
 void mg_timer_init(struct mg_timer *, int ms, int, void (*fn)(void *), void *);
 void mg_timer_free(struct mg_timer *);
-void mg_timer_poll(unsigned long uptime_ms);
+void mg_timer_poll(unsigned long long uptime_ms);
 
 
 
@@ -444,7 +444,7 @@ int mg_asprintf(char **buf, size_t size, const char *fmt, ...);
 int mg_vasprintf(char **buf, size_t size, const char *fmt, va_list ap);
 int64_t mg_to64(struct mg_str str);
 double mg_time(void);
-unsigned long mg_millis(void);
+unsigned long long mg_millis(void);
 void mg_usleep(unsigned long usecs);
 
 #if MG_ENABLE_FS
